fix(replays): moves event tracking out of components
