+++ static/app/components/replays/replayController.tsx
@@ -18,15 +18,11 @@ import {
   IconSettings,
 } from 'sentry/icons';
 import {t} from 'sentry/locale';
-import ConfigStore from 'sentry/stores/configStore';
-import {useLegacyStore} from 'sentry/stores/useLegacyStore';
 import space from 'sentry/styles/space';
 import {SelectValue} from 'sentry/types';
 import {BreadcrumbType} from 'sentry/types/breadcrumbs';
-import trackAdvancedAnalyticsEvent from 'sentry/utils/analytics/trackAdvancedAnalyticsEvent';
 import {getNextBreadcrumb} from 'sentry/utils/replays/getBreadcrumb';
 import useFullscreen from 'sentry/utils/replays/hooks/useFullscreen';
-import useOrganization from 'sentry/utils/useOrganization';
 
 const SECOND = 1000;
 
@@ -44,9 +40,6 @@ interface Props {
 }
 
 function ReplayPlayPauseBar({isCompact}: {isCompact: boolean}) {
-  const config = useLegacyStore(ConfigStore);
-  const organization = useOrganization();
-
   const {
     currentTime,
     isFinished,
@@ -56,15 +49,6 @@ function ReplayPlayPauseBar({isCompact}: {isCompact: boolean}) {
     setCurrentTime,
     togglePlayPause,
   } = useReplayContext();
-
-  const handlePlayToggle = () => {
-    trackAdvancedAnalyticsEvent('replay.play-pause', {
-      organization,
-      user_email: config.user.email,
-      play: !isPlaying,
-    });
-    togglePlayPause(!isPlaying);
-  };
 
   return (
     <ButtonBar merged>
@@ -90,7 +74,7 @@ function ReplayPlayPauseBar({isCompact}: {isCompact: boolean}) {
           size="sm"
           title={isPlaying ? t('Pause') : t('Play')}
           icon={isPlaying ? <IconPause size="sm" /> : <IconPlay size="sm" />}
-          onClick={handlePlayToggle}
+          onClick={() => togglePlayPause(!isPlaying)}
           aria-label={isPlaying ? t('Pause') : t('Play')}
         />
       )}
@@ -186,20 +170,9 @@ const ReplayControls = ({
   toggleFullscreen = () => {},
   speedOptions = [0.1, 0.25, 0.5, 1, 2, 4],
 }: Props) => {
-  const config = useLegacyStore(ConfigStore);
-  const organization = useOrganization();
   const barRef = useRef<HTMLDivElement>(null);
   const [compactLevel, setCompactLevel] = useState(0);
   const {isFullscreen} = useFullscreen();
-
-  const handleFullscreenToggle = () => {
-    trackAdvancedAnalyticsEvent('replay.toggle-fullscreen', {
-      organization,
-      user_email: config.user.email,
-      fullscreen: !isFullscreen,
-    });
-    toggleFullscreen();
-  };
 
   const updateCompactLevel = useCallback(() => {
     const {width} = barRef.current?.getBoundingClientRect() ?? {width: 500};
@@ -228,7 +201,7 @@ const ReplayControls = ({
         title={isFullscreen ? t('Exit full screen') : t('Enter full screen')}
         aria-label={isFullscreen ? t('Exit full screen') : t('Enter full screen')}
         icon={isFullscreen ? <IconContract size="sm" /> : <IconExpand size="sm" />}
-        onClick={handleFullscreenToggle}
+        onClick={toggleFullscreen}
       />
     </ButtonGrid>
   );
